configs/requirements.txt
absl-py==2.1.0
addict==2.4.0
affine==2.4.0
aiohttp==3.9.3
aiosignal==1.3.1
annotated-types==0.6.0
ansible-cmdb==1.31
anyio==4.4.0
appdirs==1.4.4
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
asttokens==2.4.1
async-lru==2.0.4
async-timeout==4.0.3
attrs==23.2.0
autocommand==2.2.2
Babel==2.15.0
backcall==0.2.0
beautifulsoup4==4.12.3
bleach==6.1.0
boto3==1.34.158
botocore==1.34.158
bravado==11.0.3
bravado-core==6.1.1
cachetools==5.3.3
certifi==2024.2.2
cffi==1.17.0
charset-normalizer==3.3.2
cheroot==10.0.0
CherryPy==18.9.0
click==8.1.7
click-plugins==1.1.1
cligj==0.7.2
comm==0.2.2
contourpy==1.1.1
cycler==0.12.1
debugpy==1.8.2
decorator==5.1.1
defusedxml==0.7.1
-e git+git@github.com:marieangeboum/dl_toolbox.git@0b3b778d34298adf714fa26bf96f5608a6fdc075#egg=dl_toolbox
docker-pycreds==0.4.0
easydict==1.13
efficientnet-pytorch==0.7.1
einops==0.7.0
exceptiongroup==1.2.2
executing==2.0.1
fastjsonschema==2.20.0
filelock==3.13.1
fiona==1.9.6
fonttools==4.49.0
frozenlist==1.4.1
fsspec==2024.2.0
future==1.0.0
geopandas==0.13.2
gitdb==4.0.11
GitPython==3.1.42
google-auth==2.28.2
google-auth-oauthlib==1.0.0
grpcio==1.62.1
h11==0.14.0
hostlist==1.4.8
httpcore==1.0.5
httpx==0.27.0
huggingface-hub==0.24.2
idna==3.6
imagesize==1.4.1
importlib-metadata==7.0.1
importlib-resources==6.1.2
inflect==7.0.0
iniconfig==2.0.0
ipykernel==6.29.5
ipython==8.12.3
ipywidgets==8.1.3
jaraco.collections==5.0.0
jaraco.context==4.3.0
jaraco.functools==4.0.0
jaraco.text==3.12.0
jedi==0.19.1
Jinja2==3.1.3
jmespath==1.0.1
joblib==1.4.2
json5==0.9.25
jsonref==1.1.0
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
jsonxs==0.6
jupyter==1.0.0
jupyter-client==8.6.2
jupyter-console==6.6.3
jupyter-core==5.7.2
jupyter-events==0.10.0
jupyter-lsp==2.2.5
jupyter-server==2.14.2
jupyter-server-terminals==0.5.3
jupyterlab==4.2.4
jupyterlab-pygments==0.3.0
jupyterlab-server==2.27.3
jupyterlab-widgets==3.0.11
kiwisolver==1.4.5
lightning-utilities==0.10.1
Mako==1.3.2
Markdown==3.5.2
MarkupSafe==2.1.5
matplotlib==3.7.5
matplotlib-inline==0.1.7
mistune==3.0.2
mmcv==1.3.8
mmsegmentation==0.14.1
monotonic==1.6
more-itertools==10.2.0
mpmath==1.3.0
msgpack==1.0.8
multidict==6.0.5
munch==4.0.0
nbclient==0.10.0
nbconvert==7.16.4
nbformat==5.10.4
neptune==1.10.4
neptune-pytorch==2.0.0
nest-asyncio==1.6.0
networkx==3.1
notebook==7.2.1
notebook-shim==0.2.4
numpy==1.24.4
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.19.3
nvidia-nvjitlink-cu12==12.3.101
nvidia-nvtx-cu12==12.1.105
oauthlib==3.2.2
opencv-python==4.9.0.80
overrides==7.7.0
packaging==23.2
pandas==2.0.3
pandocfilters==1.5.1
parso==0.8.4
pexpect==4.9.0
pickleshare==0.7.5
pillow==10.2.0
pkgutil-resolve-name==1.3.10
platformdirs==4.2.0
plotly==5.20.0
pluggy==1.5.0
portend==3.2.0
pretrainedmodels==0.7.4
prettytable==3.10.0
prometheus-client==0.20.0
prompt-toolkit==3.0.47
protobuf==4.25.3
psutil==5.9.8
ptyprocess==0.7.0
pure-eval==0.2.3
pyasn1==0.5.1
pyasn1-modules==0.3.0
pycparser==2.22
pydantic==2.6.3
pydantic-core==2.16.3
pygments==2.18.0
PyJWT==2.9.0
pyparsing==3.1.1
pyproj==3.5.0
pytest==8.3.2
python-dateutil==2.9.0.post0
python-hostlist==1.23.0
python-json-logger==2.0.7
pytorch-lightning==2.2.1
pytz==2024.1
PyYAML==6.0.1
pyzmq==26.0.3
qtconsole==5.5.2
QtPy==2.4.1
rasterio==1.3.9
referencing==0.35.1
regex==2024.7.24
requests==2.31.0
requests-oauthlib==1.4.0
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rpds-py==0.20.0
rsa==4.9
s3transfer==0.10.2
safetensors==0.4.2
scikit-learn==1.3.2
scipy==1.10.1
seaborn==0.13.2
# Editable install with no version control (segment-anything==1.0)
-e /d/maboum/dil_seg/lib/python3.8/site-packages
segmentation-models-pytorch==0.3.3
Send2Trash==1.8.3
sentry-sdk==1.42.0
setproctitle==1.3.3
shapely==2.0.5
simplejson==3.19.2
six==1.16.0
smmap==5.0.1
sniffio==1.3.1
snuggs==1.4.7
soupsieve==2.5
stack-data==0.6.3
swagger-spec-validator==3.0.4
sympy==1.12
tabulate==0.9.0
tempora==5.5.1
tenacity==8.2.3
tensorboard==2.14.0
tensorboard-data-server==0.7.2
termcolor==2.4.0
terminado==0.18.1
threadpoolctl==3.5.0
timm==0.9.2
tinycss2==1.3.0
tokenizers==0.19.1
tomli==2.0.1
torch==2.2.1
torchaudio==2.4.0+cu121
torchmetrics==1.3.1
torchsummary==1.5.1
torchvision==0.17.1
tornado==6.4.1
tqdm==4.66.2
traitlets==5.14.3
transformers==4.43.3
triton==2.2.0
typing==3.7.4.3
typing-extensions==4.10.0
tzdata==2024.1
urllib3==2.2.1
ushlex==0.99.1
wandb==0.16.4
wcwidth==0.2.13
webencodings==0.5.1
websocket-client==1.8.0
werkzeug==3.0.1
widgetsnbextension==4.0.11
yacs==0.1.8
yapf==0.40.2
yarl==1.9.4
zc.lockfile==3.0.post1
zipp==3.17.0
